move reportincludes into its own file
