+++ src/ReportIncludes.hpp
@@ -1,0 +1,4 @@
+#pragma once
+struct Driver;
+
+bool ReportIncludes(Driver *self);
